+++ Chatcraft/packages.config
@@ -2,5 +2,5 @@
 <packages>
   <package id="Microsoft.AspNet.WebApi.Client" version="5.2.3" targetFramework="net452" />
   <package id="Newtonsoft.Json" version="9.0.1" targetFramework="net452" />
-  <package id="Telegram.Bot" version="10.4.0" targetFramework="net452" />
+  <package id="Telegram.Bot" version="10.4.0" targetFramework="net452" requireReinstallation="true" />
 </packages>
